the getDuration() in Service should return int instead of double

+++ IsotopeCR-Backend/src/main/java/ca/mcgill/ecse321/model/Service.java
@@ -49,7 +49,7 @@ public class Service
   
   private int duration;
   
-  public double getDuration() {
+  public int getDuration() {
 	  return this.duration;
   }
